LPD-17767 Rename
