+++ modules/dxp/apps/commerce/commerce-machine-learning-impl/src/main/java/com/liferay/commerce/machine/learning/internal/search/constants/IndexNamePatterns.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.commerce.machine.learning.internal.search.api;
+package com.liferay.commerce.machine.learning.internal.search.constants;
 
 import com.liferay.portal.search.index.IndexNameBuilder;
